requirements.txt
streamlit>=1.30.0
numpy<2.0.0
pydantic>=1.10.7
scikit-learn>=1.3.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
langchain-chroma>=0.1.0
langchain-groq>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.2.0
wikipedia>=1.4.0
python-dotenv>=1.0.0
unstructured>=0.10.0
pypdf>=3.0.0
docx2txt>=0.8
requests>=2.31.0
toml
